requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-Session==0.5.0
Flask-WTF==1.1.1
email_validator==2.0.0
python-dotenv==1.0.0
requests==2.28.2
openai==1.5.0
Werkzeug==2.2.3
Markdown==3.4.3
Pillow==9.5.0
gunicorn==20.1.0
watchdog==3.0.0
pytest==7.3.1
pytest-flask==1.2.0
# Dependências do PostgreSQL
psycopg2-binary==2.9.9
# Cliente Supabase (opcional)
supabase-py==2.3.0
